requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-redis
redis
apache-airflow-providers-common-sql
flask
psycopg2-binary
